requirements.txt
anyio==4.0.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.0.5
wrapt==1.15.0
